+++ Assets/_Project/Scripts/Core/GameSceneId.cs
@@ -6,6 +6,7 @@ namespace PawsAndLoot.Core
         Game = 1,
         Result = 2,
         TechnicalTest = 3,
-        BlenderTechnicalTest = 4
+        BlenderTechnicalTest = 4,
+        VoiceTechnicalTest = 5
     }
 }